Add missing add_definition() calls.
For some unknown reason, while the svn to git conversion work was going
on in branches/work/kdeutils-build-standalone, the lines adding
${QT_DEFINITIONS} and ${KDE4_DEFINITIONS} to the definitions passed to
the compiler end up not being included in each application's
CMakeLists.txt.

This was not a problem during the 4.7 series, as Dirk still packaged
kdeutils as a single tarball, and the top-level CMakeLists.txt used did
include these lines.

The issue came to attention during the 4.8 beta/RC phase, when 32-bit
Linux users had problems creating tarballs with Ark: libarchive is
usually built in those systems with -D_FILE_OFFSET_BITS=64. This is set
in KDE4_DEFINITIONS, which was not being used to build Ark. struct stat
then had different sizes for Ark and libarchive, which caused all sorts
of trouble.